+++ run.sh
@@ -7,8 +7,8 @@ export DB_BACKUP_STRATEGY=0 # WARNING: this must be 0/1 bc it is used as index, 
 export JWT_SECRET_KEY=testsecretkey
 
 if [[ ! -v TEST_BACKUP ]]; then
-    ./server.sh
-    # uvicorn main:app --reload
+    # ./server.sh
+    uvicorn main:app --reload
 else
     echo "[*] Generating random data and testing backup"
     python3 ./generate_random_db_content.py && python3 ./db_backup.py
