+++ scripts/v1/admin_handling.py
@@ -5,12 +5,14 @@ from fastapi import HTTPException
 
 from db.models.admins import DBAdmins
 from db.singleton_handler import global_db_handler
-from pydantic_models.admin import AdminModel, DetailedAdminModel
-from scripts.shared.security import pwd_context
+from pydantic_models.admin import (AdminModel, DetailedAdminModel,
+                                   FullAdminModel)
+from scripts.shared.security import hash_pwd
 
 
 def get_all_admins() -> List[AdminModel]:
     "Gets each admin from db and converts it into safer model to respond with"
+    # TODO: logs
     result: List[AdminModel] = list()
     db_handler = global_db_handler()
 
@@ -25,6 +27,7 @@ def get_all_admins() -> List[AdminModel]:
 
 def add_admin_to_db(admin_credentials: DetailedAdminModel) -> None:
     "Adds an admin to the db with the original password additionally hashed"
+    # TODO: logs
     db_handler = global_db_handler()
 
     with db_handler.session() as session:
@@ -40,8 +43,22 @@ def add_admin_to_db(admin_credentials: DetailedAdminModel) -> None:
             DBAdmins.convertFromPydanticObject(
                 admin_credentials,
                 exclude = set("password"),
-                include = {"password": pwd_context.hash(admin_credentials.password)}
+                include = {"password": hash_pwd(admin_credentials.password)}
             )
         )
 
         session.commit()
+
+def change_admin_in_db(updated_admin_model: FullAdminModel) -> None:
+    # TODO: logs
+    ...
+
+def delete_admin_from_db(admin_id: int) -> None:
+    # maybe twofactor ?
+    # TODO: logs
+    ...
+
+def change_webhook_url_for(admin_id: int, new_webhook_url: str) -> None:
+    # TODO: definitely two factor
+    # TODO: logs
+    ...
